universe: move sync specific structs from interface.go -> syncer.go
