+++ universe/interface.go
@@ -303,47 +303,6 @@ func (s *ServerAddr) HostStr() string {
 	return s.addrStr
 }
 
-// SyncType is an enum that describes the type of sync that should be performed
-// between a local and remote universe.
-type SyncType uint8
-
-const (
-	// SyncIssuance is a sync that will only sync new asset issuance events.
-	SyncIssuance SyncType = iota
-
-	// SyncFull is a sync that will sync all the assets in the universe.
-	SyncFull
-)
-
-// String returns a human-readable string representation of the sync type.
-func (s SyncType) String() string {
-	switch s {
-	case SyncIssuance:
-		return "issuance"
-	case SyncFull:
-		return "full"
-	default:
-		return fmt.Sprintf("unknown(%v)", int(s))
-	}
-}
-
-// AssetSyncDiff is the result of a success Universe sync. The diff contains the
-// Universe root, and the set of assets that were added to the Universe.
-type AssetSyncDiff struct {
-	// OldUniverseRoot is the root of the universe before the sync.
-	OldUniverseRoot BaseRoot
-
-	// NewUniverseRoot is the new root of the Universe after the sync.
-	NewUniverseRoot BaseRoot
-
-	// NewAssetLeaves is the set of new leaf proofs that were added to the
-	// Universe.
-	NewLeafProofs []*MintingLeaf
-
-	// TODO(roasbeef): ability to return if things failed?
-	//  * can used a sealed interface to return the error
-}
-
 // Syncer is used to synchronize the state of two Universe instances: a local
 // instance and a remote instance. As a Universe is a tree based structure,
 // tree based bisection can be used to find the point of divergence with
@@ -449,91 +408,6 @@ type FederationLog interface {
 	// LogNewSyncs logs a new sync event for each server. This can be used
 	// to keep track of the last time we synced with a remote server.
 	LogNewSyncs(ctx context.Context, addrs ...ServerAddr) error
-}
-
-// SyncStatsSort is an enum used to specify the sort order of the returned sync
-// stats.
-type SyncStatsSort uint8
-
-const (
-	// SortByNone is a sentinel value that indicates that no sorting should
-	// be done.
-	SortByNone SyncStatsSort = iota
-
-	// SortByAssetName sorts the returned stats by the asset name.
-	SortByAssetName
-
-	// SortByAssetType sorts the returned stats by the asset type.
-	SortByAssetType
-
-	// SortByAssetID sorts the returned stats by the asset ID.
-	SortByAssetID
-)
-
-// SyncStatsQuery packages a set of query parameters to retrieve stats related
-// to the sync activity of a given Universe. Any of the filters can be
-// specified, however only a single sort by value should be specified. The
-// offset and limit fields can be used to implement pagination.
-type SyncStatsQuery struct {
-	// AssetNameFilter can be used to filter for stats for a given asset name.
-	AssetNameFilter string
-
-	// AssetIDFilter can be used to filter for stats for a given asset ID.
-	AssetIDFilter asset.ID
-
-	// AssetTypeFilter can be used to filter for stats for a given asset
-	// type.
-	AssetTypeFilter *asset.Type
-
-	// SortBy is the sort order to use when returning the stats.
-	SortBy SyncStatsSort
-
-	// Offset is the offset to use when returning the stats. This can be
-	// used to paginate the response.
-	Offset int
-
-	// Limit is the maximum number of stats to return. This can be used to
-	// paginate the response.
-	Limit int
-}
-
-// AssetSyncSnapshot is a snapshot of the sync activity for a given asset.
-type AssetSyncSnapshot struct {
-	// AssetID is the ID of the asset.
-	AssetID asset.ID
-
-	// AssetName is the name of the asset.
-	AssetName string
-
-	// AssetType is the type of the asset.
-	AssetType asset.Type
-
-	// TotalSupply is the total supply of the asset.
-	TotalSupply uint64
-
-	// GenesisHeight is the height of the block that the asset was created
-	// in.
-	GenesisHeight uint32
-
-	// TotalSyncs is the total number of syncs that have been performed for
-	// the target asset.
-	TotalSyncs uint64
-
-	// TotalProofs is the total number of proofs that have been inserted
-	// for the asset.
-	TotalProofs uint64
-
-	// TODO(roasbeef): add last sync?
-}
-
-// AssetSyncStats is the response to a SyncStatsQuery request. It contains the
-// original query, and the set of sync stats generated by the query.
-type AssetSyncStats struct {
-	// Query is the original query that was used to generate the stats.
-	Query SyncStatsQuery
-
-	// SyncStats is the set of sync stats generated by the query.
-	SyncStats []AssetSyncSnapshot
 }
 
 // AggregateStats is a set of aggregate stats for a given Universe.
